Clean up and format ObjectInfo.c

+++ Headers/GliderProtos.h
@@ -335,7 +335,7 @@ void HiliteAllObjects (void);
 void GoToObjectInRoom (SInt16, SInt16, SInt16);
 void GoToObjectInRoomNum (SInt16, SInt16);
 
-void GetObjectName (wchar_t *, int, SInt16);			// --- ObjectInfo.c
+void GetObjectName (wchar_t *, size_t, SInt16);			// --- ObjectInfo.c
 void DoObjectInfo (HWND);
 
 void GetObjectRect (objectPtr, Rect *);					// --- ObjectRects.c
